CrmConsig/requirements.txt
apturl==0.5.2
asgiref==3.2.3
backcall==0.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==4.0.2
coverage==5.0.3
crayons==0.3.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.3
django-cors-headers==3.7.0
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-filter==2.4.0
django-recaptcha==2.0.6
django-widget-tweaks==1.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
Faker==4.0.0
fasteners==0.14.1
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
mysql-connector-python==8.0.19
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycups==1.9.73
Pygments==2.5.2
PyGObject==3.36.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
sentry-sdk==0.14.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.0
systemd-python==234
testfixtures==6.11.0
text-unidecode==1.3
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.1.8
webdriver-manager==2.3.0
xkit==0.0.0
